webprox-avif: make the workspace compile and pass fmt + clippy
The tree did not build.  Six use-statements sat inside encode_frame in
cloud-render-srv (two of them duplicated), so rav1e's Context, Config,
ChromaSampling and SpeedSettings were not in scope at module level, where the
struct definition and new() need them.  manager.rs had the same problem with
SpatialMetadata and VideoFrame.  All of them are hoisted to the top of their
file now.

macroquad-client defined ClientCli twice, once in main.rs and once in
integration_test/cli.rs, so parse_integration_test_args could not be called
from main.rs (two distinct types with the same name).  The duplicate is gone;
cli.rs now takes crate::ClientCli.  test_runner.rs was missing
tokio_stream::StreamExt for stream.next(), stream_handler.rs was missing warn
from its log import.

Removed a dead local EncoderConfig in rav1e_enc.rs.  It was never constructed
and shadowed rav1e::prelude::EncoderConfig, which is why the rav1e one had to
be spelled out with its full path.

The rest is clippy -D warnings and cargo fmt:

* field_reassign_with_default: build EncoderConfig and the test ServerConfig as
  struct literals with ..Default::default()
* chunks_exact with a constant size -> as_chunks::<N>()
* redundant_pattern_matching, needless_borrow, unnecessary_cast,
  redundant_field_names, needless_return, unused_unit-style returns
* while_let_loop in the integration-test event loop
* unwrap after is_none in the render loop -> match on texture.as_mut()
* type_complexity in the AV1 decoder -> DecodedFrame type alias
* the Result of the integration test run was dropped; it is logged now
* dead-code scaffolding that is deliberately kept (save_image_as_png,
  check_link_clicks, TestImageData::rgba_data) is marked #[allow(dead_code)]
  with a note on why
* clippy::result_large_err comes from the tonic/prost generated service traits
  and is allowed for that module only

This commit also lands the previously staged but uncommitted refactor of this
example; the fixes above depend on it, so splitting it out would have produced
an intermediate commit that does not compile.

Verified: cargo fmt --all -- --check, cargo clippy --all-targets
--all-features -- -D warnings and cargo build --all-targets --all-features all
succeed with no warnings.  Needs nasm/yasm and protobuf-compiler for the native
libaom and prost build scripts.

+++ examples/20_webprox_avif/macroquad-client/src/state.rs
@@ -11,8 +11,8 @@ pub struct ClientImage {
 impl From<ClientImage> for macroquad::prelude::Image {
     fn from(img: ClientImage) -> Self {
         macroquad::prelude::Image {
-            width: img.width as u16,
-            height: img.height as u16,
+            width: img.width,
+            height: img.height,
             bytes: img.bytes,
         }
     }
@@ -28,20 +28,20 @@ pub struct ClientState {
     pub packet_count: u64,
     pub last_packet_size: usize,
     pub first_frame_received: bool,
-    
+
     // AV1 decoder
     pub decoder: Option<crate::video::decoder::Av1Decoder>,
-    
+
     // Spatial metadata
     pub links: Vec<LinkBox>,
     pub doc_width: u32,
     pub doc_height: u32,
     pub page_title: String,
-    
+
     // Virtual framebuffer
     pub server_viewport_y: i32,
     pub local_scroll_y: i32,
-    
+
     // Debug parameters
     pub debug_y_offset: i32,
     pub verbose_coords: bool,
